Pointeurs ok, répartition en plusieurs fichiers, plusieurs individus dans la pop nombre à choisir dans data.h

+++ structures.h
@@ -6,7 +6,7 @@
 typedef struct {
 
     int idIndividu[NBR_INTERFACES];
-    int listeBits[NBR_APPRENANTS * NBR_INTERFACES];
+    int8_t listeBits[NBR_APPRENANTS * NBR_INTERFACES];
     int nbHeuresSemaine[NBR_INTERFACES];
 
 } Individu;
